# fix: Fix parsing of tags with major versions higher than 9