@john24rel 2022-05-25
hjsdjbsjbsdkbdbbdsbsfjbsdjbjsfbjsfbj
Search Kubernetes




J
2
1

2

3

4

5


Kubernetes



kubernetes-users

Welcome everyone! || https://k8s.dev/events/office-hours/ || Next Office Hours Session: April 21 https://www.youtube.com/watch?v=m1p2sVlQXmM || Forum: https://discuss.kubernetes.io




149,979
Tip: Try  F to search this channel



Mike Carpendale  11:59 PM
any php experts here that can help me with the formatting. Line 6 by the looks, but could be other issues :ladybug:......
PHP Parse error: syntax error, unexpected 'SELECT' (T_STRING) in /home/pureuser/db-reset.php on line 6
<?php
$postgres = 'kubectl get pods -n migrationnamespace | grep postgres | cut -d " " -f1 2>&1';
$postgres_pod = shell_exec($postgres);
echo $postgres_pod;

$list2 = 'kubectl exec -it -n migrationnamespace ' . $postgres_pod . ' -- psql -U postgres -c 'SELECT * FROM mywhales'; 2>&1';
echo "<pre>";
echo shell_exec($list2);
echo "<pre>";
?>
$postgres_pod = postgres-7957478b7d-tmw6m
when you run this :point_down:from the CLI it executes fine - just issues from php
kubectl exec -it -n migrationnamespace postgres-7957478b7d-tmw6m -- psql -U postgres -c 'SELECT * FROM mywhales';


3 replies
Last reply 7 days agoView thread


Karl Kloppenborg  12:18 AM
Wouldn't be better suited for a PHP forum?


Alok Jha  12:28 AM
joined #kubernetes-users along with 3 others.


comfuture  1:41 AM
Anyone succeeded in using letsencrypt certificates on microk8s using cert-manager.io CRD?
I got failed ClusterIssuer with message:
 Failed to register ACME account: Get “https://acme-v02.api.letsencrypt.org/directory”: dial tcp: lookup acme-v02.api.letsencrypt.org on 10.152.183.10:53: server misbehaving


5 replies
Last reply 7 days agoView thread


Jeegar Patil  1:53 AM
joined #kubernetes-users along with wabit.


Angel  2:08 AM
How to solve latency issue in k8s ?
What is the root cause for that 75 sec or 130 sec delay ?
Root cause may be Application Load Balancer, Volume, CUP Limit, ??
any ideas about how to reduce latency in k8s microservice ?



48 replies
Last reply 7 days agoView thread


Tobias Gmünder  2:10 AM
joined #kubernetes-users along with 4 others.


Philipp Kraus  2:46 AM
Hello, I try to create Java program (with the Micronaut frameworkt) to generate K8 jobs. So I would like to run a Java pod and this pod creates K8 jobs via Java Kubernetes Client. I’m using as well Maven, so I have added the dependency
    <dependencies>
        <dependency>
            <groupId>io.kubernetes</groupId>
            <artifactId>client-java</artifactId>
            <version>15.0.1</version>
        </dependency>
    </dependencies>
but during Maven-Shade plugin to build a FatJar I get a signature error
Error creating shaded jar: Problem shading JAR /myjar.jar entry META-INF/MANIFEST.MF: java.lang.SecurityException: Invalid signature file digest for Manifest main attributes 
It seems this issue relates to the K8 client library (so if I remove it, the Jar will be generated and worked well). I try to use the Maven Shade plugin as well with this setup to remove any signature files
    <build>
        <plugins>
            <plugin>
                <groupId>org.apache.maven.plugins</groupId>
                <artifactId>maven-shade-plugin</artifactId>
                <executions>
                    <execution>
                        <phase>package</phase>
                        <goals>
                            <goal>shade</goal>
                        </goals>
                        <configuration>
                            <filters>
                                <filter>
                                    <artifact>*:*</artifact>
                                    <excludes>
                                        <exclude>META-INF/*.SF</exclude>
                                        <exclude>META-INF/*.DSA</exclude>
                                        <exclude>META-INF/*.RSA</exclude>
                                    </excludes>
                                </filter>
                            </filters>
                        </configuration>
                    </execution>
                </executions>
            </plugin>
        </plugins>
    </build>
I try to figure out, if this is problem on the Kubernetes Client library or the combination of the client library and the Micronaut framework (which I use for a Kafka Consumer).
Did you have got any idea to solve / debug this issue?
Thanks a lot


Jack  3:04 AM
Hi all!! Is there any tool to customize the way of scaling of worker nodes on eks. What I meant by customized is that scaling triggers with the help of queue service like sqs.


5 replies
Last reply 7 days agoView thread


Merlijn Sebrechts  3:22 AM
joined #kubernetes-users along with 7 others.


Jakub Tobiasz  3:59 AM
Hello everyone! Is it possible to use the same dependency twice in a one chart? I’ve tried to use alias but while helm dep build it does nothing. I need to setup two nginx, each with one sidecar with PHP application.


2 replies
Last reply 7 days agoView thread


Patryk Małek  4:01 AM
joined #kubernetes-users along with 4 others.


Kay  4:10 AM
Is it possible i can create some global variable in kubernetes and substitute the value for that global variable value in a kubernetes yml file.
alb.ingress.kubernetes.io/tags: Environment=<some_global_variable_value>
---
apiVersion: networking.k8s.io/v1
kind: Ingress
metadata:
  name: monitoring-ingress
  namespace: monitoring
  annotations:
    alb.ingress.kubernetes.io/scheme: internet-facing
    alb.ingress.kubernetes.io/tags: Environment=prod,Team=dev



6 replies
Last reply 6 days agoView thread


Elad Moshe  4:20 AM
What is the best declarative approach to automatically delete objects considering that I can’t use apply --prune since it is not support in SSA because of the removal of  the last-applied-configuration annotation in SSA.


Eric Monné Mesalles  4:25 AM
joined #kubernetes-users along with 6 others.


Rohmer Sicat  5:05 AM
Hi all in Ingress Nginx Kubernetes how can i do upstream like the config below
upstream odoo {
  server 127.0.0.1:8069;
}
upstream odoochat {
  server 127.0.0.1:8072;
}
(edited)

1 reply
6 days agoView thread


matt-kasten  5:13 AM
joined #kubernetes-users along with Said Sakuh.


Timothy Harris  5:16 AM
A question about taints+tolerations and daemonsets.
We need to add a taint to some nodes to keep deployments and statefulsets away from the nodes unless they match the toleration.
But we have some core infrastructure daemonsets that must run on all nodes. Since daemonsets respect taints we must now add a toleration for every core infrastructure daemonset.
Is this correctly understood?
If so. Trying to figure out the best approach. Requiring all workloads to use anti-affinity to steer them away from the nodes. Or using the taints and adding the tolerations to the critical daemonsets. (edited) 

1 reply
6 days agoView thread


Rafał Dziwiński  5:18 AM
joined #kubernetes-users along with Robbe.

Simple PollAPP  5:28 AM
What CRI for new cluster?
:one: Docker Engine
:two: Containerd    10
@Darren L, @Azalax, @abizake, @Ryan, @Mariusz Gomse, @George Gaál, @Gabriel Jesus Cuadros Caceres, @Ciriaco Lopez Tenorio, @Thomas Belian, @Michael Hewitt
:three: CRI-O    6
@Rafał Dziwiński, @davidcurran, @Thomas Marko, @Chandrasekhar Oruganty, @Michael Trip, @Håkan Lund
Created by @Rafał Dziwiński with /poll


Rafał Dziwiński  5:28 AM
Hello, as we know started at Kubernetes 1.24 version docker shim is removed. Now, i have to create new cluster and have a dilemma what CRI should I use?. What about that? (edited) 



6 replies
Last reply 6 days agoView thread


Diego Bonfigli  5:34 AM
joined #kubernetes-users along with 2 others.


Rafał Dziwiński  6:03 AM
If Someone click containerd/crio please some word about why :slightly_smiling_face:





8 replies
Last reply 6 days agoView thread


Keval Pansuriya  6:05 AM
joined #kubernetes-users.


Martin Kollaard  6:06 AM
I wonder what the difference is between those 2.


Keval Pansuriya  6:07 AM
:wave: Hello, team!


Maggie Masala  6:12 AM
joined #kubernetes-users.


Cristiano Colangelo  6:14 AM
https://www.oreilly.com/online-learning/leveraging-katacoda-technology.html
tl;dr Kataconda is closing down (edited) 

1 reply
6 days agoView thread


stefanzahov:the_horns:  6:18 AM
joined #kubernetes-users.


santhosh  6:26 AM
Any Solution for this issue
2 files 
image.png


image.png






4 replies
Last reply 6 days agoView thread


Joen Lindgreen Wallenstrøm  6:32 AM
joined #kubernetes-users.


Chandrasekhar Oruganty  6:39 AM
Hi All,
I installed Kyverno on OpenShift 4.8 and has setup Namespace selector in config map. Autoupdatewebhooks flag is set to true and am using namespaceSelector but it looks like the namespaceSelector is not reflected in validatingWebhookConfiguration. Any thoughts?


2 replies
Last reply 6 days agoView thread


Tamas  6:39 AM
joined #kubernetes-users along with 3 others.


raymond dowden  8:04 AM
Senior full-stack engineer with 11+ years of hands-on experience in web development who specializes in JavaScript Frameworks and Libraries such as React and Vue.JS for front end and Node.js for back end.
Well known for completing projects ahead of deadlines and expected cleaned code, so Received 95% positive feedback from company for testing and flexibility.

1 reply
6 days agoView thread


Patrick Bergstrom  8:22 AM
joined #kubernetes-users along with 2 others.


Mohanraj M  9:03 AM
hi all,
i need to create the node group in private subnet, but it ask to enable the auto assign public IP.
if i enabled it create the nodes with public on private subnet.
is this expected or I need to change something ?

1 reply
6 days agoView thread


fabwer  9:29 AM
joined #kubernetes-users along with 2 others.


Katherin Miranda  9:52 AM
Hello folks,
I'm using kind and want to create my cluster on a specific kubernetes version, do you know the appropriate command for it??
Thank you


2 replies
Last reply 6 days agoView thread


fabwer  9:53 AM
Hi all!
How can I make sure that an endpoint IS a Kubernetes API (including multiple versions and flavors like k8s, openshift, ...).
Ideas in :thread:

1 reply
6 days agoView thread


João Vasques  9:57 AM
joined #kubernetes-users along with 2 others.


Arun Baskaran  10:18 AM
Hi all ! How do i choose between multiple container orchestration tools (Opensource Kubernetes , AKS , EKS, Open shift , TKG etc...).. Which one best suits for a customer



6 replies
Last reply 6 days agoView thread


Mackenzie Smith  10:43 AM
joined #kubernetes-users along with 12 others.
New


Philipp Kraus  1:05 PM
Hello, I’m searching for a minimal example of the Kubernetes client library (Java, Go or Python), which shows how to create a job for an external operator. So my use-case is that I have deployed an operator into my namespace and I would like to create a job, that triggers this operator, by a running pod via the client api. In Java I have written this code:
final V1ObjectMeta l_meta = new V1ObjectMeta();
l_meta.setName( "myjob" );
l_meta.setNamespace( Namespaces.getPodNamespace() );

final V1JobSpec l_spec = new V1JobSpec();
// here I need some help how to define the spec section int he job with individual parameters

final V1Job l_job = new V1Job();
l_job.apiVersion( "myoperator/v1" );
l_job.kind( "MyOperatorApp" );
l_job.setMetadata( l_meta );
l_job.setSpec( l_spec );


12 replies
Last reply 6 days agoView thread


Patrick  1:15 PM
joined #kubernetes-users along with Zeeshan Adnan.


Seth  1:50 PM
I’m attempting to get pod logs via k8 api (golang), and for some reason i get a response of
unable to retrieve container logs for containerd://468b6c1362d9088ada346f00ec14517fd1440827fb2736b4983fe50cc050e4c3
The issue is that this is comes back as the log itself and not an error.
I’m just kind of wondering if this is expected

1 reply
6 days agoView thread


Jacek Frydrychowicz  1:55 PM
joined #kubernetes-users along with 3 others.


Robbe Verhelst  2:56 PM
Hi everyone, we have a pod with an initcontainer that updates the configmap of a different pod. The other pod is a sort of loadbalancer that needs to know of the new pods existence by adding it in its configmap. Now when we take down the first pod, we want it to run the initcontainer again, so it can delete itself from the configmap of the loadbalancer pod. The problem is that the initcontainer works perfectly, but we now also need something like an "outit-container". How would any of you tackle this? To run a container before starting and running that container after terminating.


3 replies
Last reply 3 days agoView thread


David Ezequiel Borges  2:57 PM
joined #kubernetes-users.


Victor  3:02 PM
Hi everyone.
I’ve got a CSI Driver that doesn’t receive API calls even though the provisioner specified correctly. The pod attaching a PVC through this CSI driver is in pending state and I guess it reaches to Kube API server, but it never gets to the CSI Driver. I want to troubleshoot/debug the Kube API server to find why/where the request gets stuck - how can I do that? Where the Kube API server logs, that catch such events, are located?
Thanks!


Sipnek  3:18 PM
joined #kubernetes-users along with 2 others.


Cristiano Colangelo  3:28 PM
does anybody know if it's possible to somehow deploy a working (temporary) Kubernetes cluster using kind or something else in a gitlab job?


3 replies
Last reply 6 days agoView thread


Alexander Niebuhr  3:44 PM
Hey folks,
what underlaying linux distro is the best to use with k8s. The VMs will just run k8s. (Talos not an option, because of missing sh)
Things we care about are memory / cpu usages & up-to date packages.
What are you using / what to you thing is used by big deployments?

1 reply
6 days agoView thread


Stoimen Hristov  3:51 PM
joined #kubernetes-users along with 3 others.


Stefan Lasiewski  4:38 PM
I want to test if kubectl is able to connect to the cluster. Is there something like a kubectl ping-cluster command to provide a simple connection test, or should I instead just use something like kubectl cluster-info.


2 replies
Last reply 6 days agoView thread


Vito Botta  4:50 PM
I enabled vulnerability scanning for container images in GCP Artifact Registry and it found a whopping 400+ vulnerabilities, including 2 critical and 10ish high priority. The rest are medium and low. When building the image all OS packages get upgraded. How is it possible that it finds so many vulnerabilities? What do you do to minimize this?

1 reply
6 days agoView thread


Wei-chung Wang  5:21 PM
joined #kubernetes-users along with Salatiel Filho.


krisM  6:12 PM
does anybody know of a good tool that can render gotemplates to yaml files? I am currently using shell scripts to do this but it becomes a mess when escaping certain characters in strings, like when rendering SQL in a configmap, or regex/sed on values. Is there a tool that I can run against a gotemplate to achieve something like this? something similar to how helm works

1 reply
6 days agoView thread


xdexter  6:29 PM
Hi! I have this entry on kube-apiserver log:
I0519 18:00:33.932730       1 trace.go:205] Trace[1169215247]: "Delete" url:/api/v1/namespaces/production/secrets,user-agent:kube-controller-manager/v1.21.11 (linux/amd64) kubernetes/38d3c1f/system:serviceaccount:kube-system:namespace-controller,client:127.0.0.1,accept:application/vnd.kubernetes.protobuf,application/json,protocol:HTTP/2.0 (19-May-2022 18:00:29.299) (total time: 4632ms):
6:30
It's possible identity the source? Who executed the action?
6:30
(I don't have audit enabled)


vc  6:55 PM
joined #kubernetes-users along with 5 others.


Adam Gardner  9:33 PM
I notice that Kuberentes has a package called credentialsprovider which has a file, config.go which defines types like DockerConfigEntry, etc, all of which define the code necessary to read registry credentials from a docker-style config.json file (with some extra features, apparently?). I had hopes this would mean that the kubelet could read these credential files directly, even if not using docker as a runtime, but, when I switch to containerd as my backend, the auth information I set up is ignored.

3 replies
Last reply 3 days agoView thread


Adam Gardner  9:35 PM
Does anyone have any tips on how to set up node-level auth when using the containerd runtime? As far as I can tell, containerd itself has had 3 different ways to configure this within the runtime, and all of them are deprecated? People seem to have real trouble getting any of them work, if the github issues I've seen are any indication.

1 reply
6 days agoView thread


Hung Duong  9:42 PM
joined #kubernetes-users along with daMaestro.


daMaestro  9:46 PM
hi, i'm trying to create a dual node pool where one pool runs "everything else" and the other pool only runs pods with a specific toleration. the purpose is to have the standard services running on the "everything else" pool without any toleration so i can run whatever and then a pool that only runs something that has the toleration
9:48
i'm trying to auto-scale the toleration pool to zero when those pods are not active. i've successfully tainted the zero pool so everything is running on the nodes without taints
9:48
i'm curious, is there a way to taint a pool to not accept pods with a certain toleration? so far i've not figure a way to make that happen
9:49
the pods with the toleration are either scheduled on the tainted node pool or the "everything else" pool. i'm trying to get the tolerated pods only onto the tainted pool.
9:49
thanks for any insight on what i'm missing.


daMaestro  10:24 PM
https://kubernetes.io/docs/concepts/scheduling-eviction/assign-pod-node/#built-in-node-labels was a good recommendation to pursue
KubernetesKubernetes
Assigning Pods to Nodes
You can constrain a Pod so that it can only run on particular set of node(s). There are several ways to do this and the recommended approaches all use label selectors to facilitate the selection. Generally such constraints are unnecessary, as the scheduler will automatically do a reasonable placement (for example, spreading your Pods across nodes so as not place Pods on a node with insufficient free resources). However, there are some circumstances where you may want to control which node the Pod deploys to, for example, to ensure that a Pod ends up on a node with an SSD attached to it, or to co-locate Pods from two different services that communicate a lot into the same availability zone.


Linh Nguyen Hoang  10:30 PM
hi all, i have a deployment config rolling-update have pvc readwriteonce so when i restart deployment will assign pod to another node then new pod will be fail to mount this pvc. may i use node affinity




12 replies
Last reply 6 days agoView thread


Ashok Srirama  10:30 PM
joined #kubernetes-users along with 5 others.


SHIBASIS  1:14 AM
Hi Team, Can anyone please help me with the exact baseurl for Kubernetes repositry configuration. I am trying with the kubeadm but can't able to see any software packages in below URL.
https://packages.cloud.google.com/yum/repos/kubernetes-el7-\$basearch


4 replies
Last reply 5 days agoView thread


Saurabh  1:16 AM
joined #kubernetes-users along with 5 others.


Thomas Güttler  2:56 AM
…